extract into contract file
